+++ CHANGELOG.md
@@ -18,8 +18,6 @@
 - Deprecate `buf mod open`.
 - Delete `buf beta migrate-v1beta1` This is now replaced with `buf config migrate`.
 - Add `buf registry sdk version` to get the version of a Generated SDK for a module and plugin.
-- Add `buf beta registry label archive` and `buf beta registry label unarchive` commands for
-  managing labels on the BSR.
 - Add support for Protobuf Editions. This allows `buf` to be used with sources that use edition
   2023, instead of proto2 or proto3 syntax. This also updates the `protoc-gen-buf-breaking` and
   `protoc-gen-buf-lint` Protobuf plugins to support files that use edition 2023.
@@ -58,6 +56,8 @@
   `features.field_presence = LEGACY_REQUIRED` in Editions files. This new rule is not active by
   default in existing `v1` and `v1beta1` configurations, for backwards-compatibility reasons.
   Migrate your config to `v2` to use them.
+- Add `buf beta registry archive` and `buf beta registry unarchive` commands for managing
+  labels on the BSR.
 
 ## [v1.32.0-beta.1] - 2024-04-23